Remove broken links from reference manual
This fixes all broken links for the reference manual.  While some internal
links were slightly rewritten, many of these links leave the scope of the
Cinderella documentation scraped for CindyJS, and were simply removed.
If we ever extend the scope of that scraping, we might want to revisit this
commit here and re-introduce some of the links removed here.