Introduce CenterOfCircle helper
While CenterOfConic can handle that situation as well, having a specialized
version avoids some computation.  Furthermore, the center of a circle which
degenerated to a line is well-defined at the point at infinity perpendicular
to the line.  For the corresponding degenerate conic the center is no longer
defined in this situation.  So having a special function avoids
singularities.