Conform to recent getUserMedia constraint API
In the long run, most browsers should properly perform candidate selection
based on fitness, so that ideal values (like the huge width and height in
this case) get matched as closely as possible, leading to maximal resolution
on pretty much all realistic devices.

However, the spec only says browsers SHOULD choose candidates based on
fitness, not that they MUST do so.  And in fact Chrome currently does NOT do
so, so we need another mechanism to control resolution selection.  But the
optional key as well as the minWidth constraints are not part of the current
specification, so we should use their more recent counterparts instead.

This now seems to have the desired effect on Firefox and Chrome.