Combine different moves of the same object
If the mouse and the user input move the same object, then these moves
should be combined into a single move.  Which means that if the user moves
to a singular situation, and the script then moves to a non-singular one,
the move to that non-singuar one will be traced from the non-singular last
good position preceding the mouse move.

Or conversely, if the user moves to a non-singular situation but the script
moves to a singular one, then subsequent moves will be based of the
user-entered position, disregarding the singular position caused by the
script.

This also affects mouse-only moves: lifting the mouse finger in a singular
situation will now no longer destroy the last good situation, and will
therefore allow tracing past the critical situation when the same object is
moved again.

The main case distinction is now in traceMover.  If the mover is kept, it
will restore the last good situation and operate on that.  If the mover is
changed, it will take the current situation as good, forgetting about past
good situations and tracing problems from these.