Avoid changing position for existing non-point elements
This is a drive-by fix: if there is a duplicate name in the list of
geometric elements, setting its position only makes sense if it is a point,
or something similar which matches our assumed position format.  If the
duplicate element is e.g. a CircleMr, this might cause an exception.